requirements.txt
Django==2.1.4
pytz==2018.7
psycopg2-binary==2.7.6.1
djangorestframework==3.9.0
django-cors-headers==2.4.0
